move functionality from AnalysisConfigGroup into ControllerConfigGroup
ControllerConfigGroup already contains many of those set*Interval methods to control what output should be generated, e.g. not only events and plans-file writing, but also writing the trips.csv etc. So I decided to move the new config params here instead of a separate config group.
