+++ matsim/src/main/java/org/matsim/core/config/groups/ControllerConfigGroup.java
@@ -225,6 +225,31 @@ public final class ControllerConfigGroup extends ReflectiveConfigGroup {
         this.writeTripsInterval = writeTripsInterval;
     }
 
+	@Parameter
+	@Comment("Defines in which iterations the legHistogram analysis is run (iterationNumber % interval == 0). Use 0 to disable this analysis.")
+	private int legHistogramInterval = 1;
+
+	@Parameter
+	@Comment("Defines in which iterations the legDurations analysis is run (iterationNumber % interval == 0). Use 0 to disable this analysis.")
+	private int legDurationsInterval = 1;
+
+	public int getLegHistogramInterval() {
+		return this.legHistogramInterval;
+	}
+
+	public void setLegHistogramInterval(int legHistogramInterval) {
+		this.legHistogramInterval = legHistogramInterval;
+	}
+
+	public int getLegDurationsInterval() {
+		return this.legDurationsInterval;
+	}
+
+	public void setLegDurationsInterval(int legDurationsInterval) {
+		this.legDurationsInterval = legDurationsInterval;
+	}
+
+
 	@StringSetter( RUNID )
 	public void setRunId(final String runid) {
 		if (runid == null) {
